chore: remove legacy auth middleware
