+++ moohaar-backend/src/controllers/admin.controller.js
@@ -1,5 +1,4 @@
-export const getDashboard = (_req, res) => {
-  return res.json({ message: 'admin dashboard placeholder' });
-};
+export const getDashboard = (_req, res) =>
+  res.json({ message: 'admin dashboard placeholder' });
 
 export default { getDashboard };
