+++ moohaar-backend/src/middleware/authorizeAdmin.js
@@ -1,0 +1,6 @@
+export default (req, res, next) => {
+  if (req.user && req.user.role === 'admin') {
+    return next();
+  }
+  return res.status(403).json({ message: 'Forbidden' });
+};
