+++ moohaar-backend/src/server.js
@@ -17,7 +17,8 @@ import healthRoutes from './routes/health.routes';
 import authRoutes from './routes/auth.routes';
 import adminRoutes from './routes/admin.routes';
 import PageView from './models/pageview.model';
-import { auth, authorizeAdmin } from './middleware/auth.middleware';
+import auth from './middleware/auth';
+import authorizeAdmin from './middleware/authorizeAdmin';
 import errorHandler from './middleware/errorHandler';
 import logger from './utils/logger';
 import { initCache, getCache, setCache } from './services/cache.service';
